Switch away from OMD, and specifically for Debian hosts.
We're not moving to using OMD, hence we need a puppet module that only sets
up check-mk.  Given this module is called puppet-check_mk, and not
puppet-omd, this seems appropriate.  We're only using this on Debian, where
check-mk packages are available.  So we're removing the install_tarball
class, and references to the $filestore that was needed for that.